+++ frontend/src/metabase/public/containers/PublicOrEmbeddedQuestion/PublicOrEmbeddedQuestion.unit.spec.tsx
@@ -14,7 +14,7 @@ import {
 } from "metabase-types/api/mocks";
 import { createMockState } from "metabase-types/store/mocks";
 
-import { PublicQuestion } from "./PublicQuestion";
+import { PublicQuestion } from "./PublicOrEmbeddedQuestion";
 
 registerVisualizations();
