+++ frontend/src/metabase/public/containers/PublicOrEmbeddedQuestion/index.ts
@@ -1,0 +1,1 @@
+export { PublicQuestion } from "./PublicOrEmbeddedQuestion";
