--- frontend/src/metabase/public/containers/PublicQuestion/index.js
@@ -1,1 +1,0 @@
-export { PublicQuestion } from "./PublicQuestion";
