+++ frontend/src/metabase/routes-public.jsx
@@ -3,7 +3,7 @@ import { PublicNotFound } from "metabase/public/components/PublicNotFound";
 import PublicAction from "metabase/public/containers/PublicAction";
 import PublicApp from "metabase/public/containers/PublicApp";
 import { PublicOrEmbeddedDashboardControlled } from "metabase/public/containers/PublicOrEmbeddedDashboard";
-import { PublicQuestion } from "metabase/public/containers/PublicQuestion";
+import { PublicQuestion } from "metabase/public/containers/PublicOrEmbeddedQuestion";
 
 import { getApplicationName } from "./selectors/whitelabel";
